removing lexerContracts
