+++ src/lexer/index.ts
@@ -2,6 +2,5 @@ export * from "./comment";
 export * from "./error";
 export * from "./keywords";
 export * from "./lexer";
-export * from "./lexerContracts";
 export * from "./lexerSnapshot";
 export * from "./token";
