+++ src/lexer/lexer.ts
@@ -3,9 +3,8 @@ import { Option } from "../common/option";
 import { PartialResult, PartialResultKind } from "../common/partialResult";
 import { LexerError } from "./error";
 import { Keyword } from "./keywords";
-import { ErrorLine, LexerLineKind, LexerLineString, LexerMultilineKind, LexerRead, LexerState, TErrorLexerLine, TLexerLine, TouchedWithErrorLine, UntouchedLine } from "./lexerContracts";
 import { LexerSnapshot } from "./lexerSnapshot";
-import { LexerLinePosition, LineToken, LineTokenKind } from "./token";
+import { LineToken, LineTokenKind } from "./token";
 
 // the lexer is
 //  * functional
@@ -17,6 +16,91 @@ import { LexerLinePosition, LineToken, LineTokenKind } from "./token";
 // Lexer.snapshot creates a frozen copy of a lexer state
 
 export namespace Lexer {
+
+    export type TLexerLineExceptUntouched = Exclude<TLexerLine, UntouchedLine>;
+
+    export type TLexerLine = (
+        | TouchedLine
+        | UntouchedLine
+        | TouchedWithErrorLine
+        | ErrorLine
+    )
+
+    export type TErrorLexerLine = (
+        | ErrorLine
+        | TouchedWithErrorLine
+    )
+
+    export interface LexerState {
+        readonly lines: ReadonlyArray<TLexerLine>,
+        readonly lineSeparator: string,
+    }
+
+    export const enum LexerMultilineKind {
+        Comment = "Comment",
+        Default = "Default",
+        QuotedIdentifier = "QuotedIdentifier",
+        String = "String",
+    }
+
+    export interface ILexerLine {
+        readonly kind: LexerLineKind,
+        readonly lineString: LexerLineString,               // text representation for the line
+        readonly numberOfActions: number,                   // allows for quick LexerLine equality comparisons
+        readonly lineNumber: number,                        // what line number is this
+        readonly tokens: ReadonlyArray<LineToken>,          // LineTokens lexed so far
+        readonly multilineKind: LexerMultilineKind,
+    }
+
+    export interface LexerLineString {
+        readonly text: string,
+        readonly graphemes: ReadonlyArray<string>,
+        readonly textIndex2GraphemeIndex: { [textIndex: number]: number; }
+        readonly graphemeIndex2TextIndex: { [graphemeIndex: number]: number; }
+    }
+
+    export const enum LexerLineKind {
+        Error = "Error",
+        Touched = "Touched",
+        TouchedWithError = "TouchedWithError",
+        Untouched = "Untouched",
+    }
+
+    export interface ErrorLine extends ILexerLine {
+        readonly kind: LexerLineKind.Error,
+        readonly error: LexerError.TLexerError,
+    }
+
+    // the last read attempt succeeded without encountering an error.
+    // possible that only whitespace was consumed.
+    export interface TouchedLine extends ILexerLine {
+        readonly kind: LexerLineKind.Touched,
+        readonly lastRead: LexerRead,
+    }
+
+    // the last read attempt read at least one token or comment before encountering an error
+    export interface TouchedWithErrorLine extends ILexerLine {
+        readonly kind: LexerLineKind.TouchedWithError,
+        readonly error: LexerError.TLexerError,
+        readonly lastRead: LexerRead,
+    }
+
+    // a call to appendtToDocument clears existing state marking it ready to be lexed
+    export interface UntouchedLine extends ILexerLine {
+        readonly kind: LexerLineKind.Untouched,
+        readonly maybeLastRead: Option<LexerRead>,
+    }
+
+    export interface LexerLinePosition {
+        readonly textIndex: number,
+        readonly columnNumber: number,
+    }
+
+    export interface LexerRead {
+        readonly tokens: ReadonlyArray<LineToken>,
+        readonly multilineKind: LexerMultilineKind,
+    }
+
 
     export function lexerLineStringFrom(text: string): LexerLineString {
         const graphemes = StringHelpers.graphemeSplitter.splitGraphemes(text);
@@ -105,6 +189,11 @@ export namespace Lexer {
             lines: state.lines.concat(newLine),
         };
         return tokenizeLine(newState, newState.lines.length - 1);
+    }
+
+    interface ReadOutcome {
+        readonly token: LineToken,
+        readonly multilineKind: LexerMultilineKind,
     }
 
     function tokenizeLine(state: LexerState, lineNumber: number): LexerState {
@@ -861,16 +950,4 @@ export namespace Lexer {
         return new LexerError.UnexpectedReadError(LexerLinePosition);
     }
 
-    // function unterminatedStringError(
-    //     text: string,
-    //     textIndex: number,
-    // ): LexerError.UnterminatedStringError {
-    //     const LexerLinePosition = StringHelpers.graphemePositionAt(text, textIndex);
-    //     return new LexerError.UnterminatedStringError(LexerLinePosition);
-    // }
-}
-
-interface ReadOutcome {
-    readonly token: LineToken,
-    readonly multilineKind: LexerMultilineKind,
 }
